Local probe on timeout
