+++ pipeline/input_copy.go
@@ -1,27 +1,22 @@
-package clients
+package pipeline
 
 import (
 	"context"
+	"errors"
 	"fmt"
-	"io"
 	"net/http"
 	"net/url"
+	"os/exec"
 	"path"
 	"strings"
-	"time"
+	"syscall"
 
-	"github.com/cenkalti/backoff/v4"
-	"github.com/hashicorp/go-retryablehttp"
+	"github.com/livepeer/catalyst-api/clients"
 	"github.com/livepeer/catalyst-api/config"
 	"github.com/livepeer/catalyst-api/crypto"
-	xerrors "github.com/livepeer/catalyst-api/errors"
 	"github.com/livepeer/catalyst-api/log"
 	"github.com/livepeer/catalyst-api/video"
-	"github.com/livepeer/go-tools/drivers"
 )
-
-const MaxCopyFileDuration = 2 * time.Hour
-const PresignDuration = 24 * time.Hour
 
 // These probe errors were found in the past on mist recordings but still process fine so we are ignoring them
 var ignoreProbeErrs = []string{
@@ -30,11 +25,11 @@ var ignoreProbeErrs = []string{
 }
 
 type InputCopier interface {
-	CopyInputToS3(requestID string, inputFile, osTransferURL *url.URL, decryptor *crypto.DecryptionKeys) (video.InputVideo, string, error)
+	CopyInputToS3(job *UploadJobPayload, inputFile, osTransferURL *url.URL, decryptor *crypto.DecryptionKeys) (video.InputVideo, error)
 }
 
 type InputCopy struct {
-	S3    S3
+	S3    clients.S3
 	Probe video.Prober
 }
 
@@ -45,9 +40,10 @@ func NewInputCopy() *InputCopy {
 }
 
 // CopyInputToS3 copies the input video to our S3 transfer bucket and probes the file.
-func (s *InputCopy) CopyInputToS3(requestID string, inputFile, osTransferURL *url.URL, decryptor *crypto.DecryptionKeys) (inputVideoProbe video.InputVideo, signedURL string, err error) {
+func (s *InputCopy) CopyInputToS3(job *UploadJobPayload, inputFile, osTransferURL *url.URL, decryptor *crypto.DecryptionKeys) (inputVideoProbe video.InputVideo, err error) {
 	var (
-		size int64
+		size      int64
+		requestID = job.RequestID
 	)
 
 	size, err = CopyAllInputFiles(requestID, inputFile, osTransferURL, decryptor)
@@ -57,17 +53,33 @@ func (s *InputCopy) CopyInputToS3(requestID string, inputFile, osTransferURL *ur
 	}
 	log.Log(requestID, "Copied", "bytes", size, "source", inputFile.String(), "dest", osTransferURL.String())
 
-	signedURL, err = getSignedURL(osTransferURL)
+	job.SignedSourceURL, err = getSignedURL(osTransferURL)
 	if err != nil {
 		return
 	}
 
 	log.Log(requestID, "starting probe", "source", inputFile.String(), "dest", osTransferURL.String())
-	inputVideoProbe, err = s.Probe.ProbeFile(requestID, signedURL)
+	inputVideoProbe, err = s.Probe.ProbeFile(requestID, job.SignedSourceURL)
 	if err != nil {
-		log.Log(requestID, "probe failed", "err", err, "source", inputFile.String(), "dest", osTransferURL.String())
-		err = fmt.Errorf("error probing MP4 input file from S3: %w", err)
-		return
+		var exitErr *exec.ExitError
+		if errors.As(err, &exitErr) && syscall.Signal(exitErr.ExitCode()) == syscall.SIGKILL ||
+			errors.Is(err, context.Canceled) {
+			// probe timed out, copy to local disk and try again
+			if err = job.CopySourceToDisk(); err != nil {
+				return
+			}
+			log.LogNoRequestID("probing local", job.localSourceFile.Name())
+			inputVideoProbe, err = s.Probe.ProbeFile(requestID, job.localSourceFile.Name())
+			if err != nil {
+				log.Log(requestID, "probe failed", "err", err, "source", inputFile.String(), "dest", osTransferURL.String())
+				err = fmt.Errorf("error probing MP4 input file from S3: %w", err)
+				return
+			}
+		} else {
+			log.Log(requestID, "probe failed", "err", err, "source", inputFile.String(), "dest", osTransferURL.String())
+			err = fmt.Errorf("error probing MP4 input file from S3: %w", err)
+			return
+		}
 	}
 	log.Log(requestID, "probe succeeded", "source", inputFile.String(), "dest", osTransferURL.String())
 	videoTrack, err := inputVideoProbe.GetTrack(video.TrackTypeVideo)
@@ -105,15 +117,7 @@ func getSignedURL(osTransferURL *url.URL) (string, error) {
 		return signedURL, nil
 	}
 
-	return SignURL(osTransferURL)
-}
-
-func IsHLSInput(inputFile *url.URL) bool {
-	ext := strings.LastIndex(inputFile.Path, ".")
-	if ext == -1 {
-		return false
-	}
-	return inputFile.Path[ext:] == ".m3u8"
+	return clients.SignURL(osTransferURL)
 }
 
 // Given a source manifest URL (e.g. https://storage.googleapis.com/foo/bar/output.m3u8) and
@@ -155,16 +159,16 @@ func getSegmentTransferLocation(srcManifestUrl, dstTransferUrl *url.URL, srcSegm
 // it will copy just the single video file for MP4/MOV input
 func CopyAllInputFiles(requestID string, srcInputUrl, dstOutputUrl *url.URL, decryptor *crypto.DecryptionKeys) (size int64, err error) {
 	fileList := make(map[string]string)
-	if IsHLSInput(srcInputUrl) {
+	if clients.IsHLSInput(srcInputUrl) {
 		// Download the m3u8 manifest using the input url
-		playlist, err := DownloadRenditionManifest(requestID, srcInputUrl.String())
+		playlist, err := clients.DownloadRenditionManifest(requestID, srcInputUrl.String())
 		if err != nil {
 			return 0, fmt.Errorf("error downloading HLS input manifest: %s", err)
 		}
 		// Save the mapping between the input m3u8 manifest file to its corresponding OS-transfer destination url
 		fileList[srcInputUrl.String()] = dstOutputUrl.String()
 		// Now get a list of the OS-compatible segment URLs from the input manifest file
-		sourceSegmentUrls, err := GetSourceSegmentURLs(srcInputUrl.String(), playlist)
+		sourceSegmentUrls, err := clients.GetSourceSegmentURLs(srcInputUrl.String(), playlist)
 		if err != nil {
 			return 0, fmt.Errorf("error generating source segment URLs for HLS input manifest: %s", err)
 		}
@@ -185,7 +189,7 @@ func CopyAllInputFiles(requestID string, srcInputUrl, dstOutputUrl *url.URL, dec
 	for inFile, outFile := range fileList {
 		log.Log(requestID, "Copying input file to S3", "source", inFile, "dest", outFile)
 
-		size, err = CopyFileWithDecryption(context.Background(), inFile, outFile, "", requestID, decryptor)
+		size, err = clients.CopyFileWithDecryption(context.Background(), inFile, outFile, "", requestID, decryptor)
 
 		if err != nil {
 			err = fmt.Errorf("error copying input file to S3: %w", err)
@@ -200,103 +204,8 @@ func CopyAllInputFiles(requestID string, srcInputUrl, dstOutputUrl *url.URL, dec
 	return size, nil
 }
 
-func isDirectUpload(inputFile *url.URL) bool {
-	return strings.HasSuffix(inputFile.Host, "storage.googleapis.com") &&
-		strings.HasPrefix(inputFile.Path, "/directUpload") &&
-		(inputFile.Scheme == "https" || inputFile.Scheme == "http")
-}
-
-func CopyFileWithDecryption(ctx context.Context, sourceURL, destOSBaseURL, filename, requestID string, decryptor *crypto.DecryptionKeys) (writtenBytes int64, err error) {
-	dStorage := NewDStorageDownload()
-	err = backoff.Retry(func() error {
-		// currently this timeout is only used for http downloads in the getFileHTTP function when it calls http.NewRequestWithContext
-		ctx, cancel := context.WithTimeout(ctx, MaxCopyFileDuration)
-		defer cancel()
-
-		byteAccWriter := ByteAccumulatorWriter{count: 0}
-		defer func() { writtenBytes = byteAccWriter.count }()
-
-		var c io.ReadCloser
-		c, err := GetFile(ctx, requestID, sourceURL, dStorage)
-
-		if err != nil {
-			return fmt.Errorf("download error: %w", err)
-		}
-
-		defer c.Close()
-
-		if decryptor != nil {
-			decryptedFile, err := crypto.DecryptAESCBC(c, decryptor.DecryptKey, decryptor.EncryptedKey)
-			if err != nil {
-				return fmt.Errorf("error decrypting file: %w", err)
-			}
-			c = io.NopCloser(decryptedFile)
-		}
-
-		content := io.TeeReader(c, &byteAccWriter)
-
-		err = UploadToOSURL(destOSBaseURL, filename, content, MaxCopyFileDuration)
-		if err != nil {
-			log.Log(requestID, "Copy attempt failed", "source", sourceURL, "dest", path.Join(destOSBaseURL, filename), "err", err)
-		}
-		return err
-	}, UploadRetryBackoff())
-	return
-}
-
-func CopyFile(ctx context.Context, sourceURL, destOSBaseURL, filename, requestID string) (writtenBytes int64, err error) {
-	return CopyFileWithDecryption(ctx, sourceURL, destOSBaseURL, filename, requestID, nil)
-}
-
-func GetFile(ctx context.Context, requestID, url string, dStorage *DStorageDownload) (io.ReadCloser, error) {
-	_, err := drivers.ParseOSURL(url, true)
-	if err == nil {
-		return DownloadOSURL(url)
-	} else if IsDStorageResource(url) && dStorage != nil {
-		return dStorage.DownloadDStorageFromGatewayList(url, requestID)
-	} else {
-		return getFileHTTP(ctx, url)
-	}
-}
-
-var retryableHttpClient = newRetryableHttpClient()
-
-func newRetryableHttpClient() *http.Client {
-	client := retryablehttp.NewClient()
-	client.RetryMax = 5                          // Retry a maximum of this+1 times
-	client.RetryWaitMin = 200 * time.Millisecond // Wait at least this long between retries
-	client.RetryWaitMax = 5 * time.Second        // Wait at most this long between retries (exponential backoff)
-	client.HTTPClient = &http.Client{
-		// Give up on requests that take more than this long - the file is probably too big for us to process locally if it takes this long
-		// or something else has gone wrong and the request is hanging
-		Timeout: MaxCopyFileDuration,
-	}
-
-	return client.StandardClient()
-}
-
-func getFileHTTP(ctx context.Context, url string) (io.ReadCloser, error) {
-	req, err := http.NewRequestWithContext(ctx, "GET", url, nil)
-	if err != nil {
-		return nil, xerrors.Unretriable(fmt.Errorf("error creating http request: %w", err))
-	}
-	resp, err := retryableHttpClient.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("error on import request: %w", err)
-	}
-	if resp.StatusCode >= 300 {
-		resp.Body.Close()
-		err := fmt.Errorf("bad status code from import request: %d %s", resp.StatusCode, resp.Status)
-		if resp.StatusCode < 500 {
-			err = xerrors.Unretriable(err)
-		}
-		return nil, err
-	}
-	return resp.Body, nil
-}
-
 type StubInputCopy struct{}
 
-func (s *StubInputCopy) CopyInputToS3(requestID string, inputFile, osTransferURL *url.URL, decryptor *crypto.DecryptionKeys) (video.InputVideo, string, error) {
-	return video.InputVideo{}, "", nil
+func (s *StubInputCopy) CopyInputToS3(job *UploadJobPayload, inputFile, osTransferURL *url.URL, decryptor *crypto.DecryptionKeys) (video.InputVideo, error) {
+	return video.InputVideo{}, nil
 }
